+++ faddr/database.py
@@ -3,109 +3,34 @@
 import ipaddress
 from datetime import datetime
 from pathlib import Path
-from typing import Dict, List
 
-from pydantic import BaseModel
-from sqlalchemy import (
-    Boolean,
-    Column,
-    ForeignKey,
-    Integer,
-    String,
-    create_engine,
-    select,
-)
-from sqlalchemy.orm import Session, declarative_base
+from sqlalchemy import create_engine, select
+from sqlalchemy.orm import Session
 
 from faddr import logger
 from faddr.exceptions import FaddrDatabaseDirError
-
-Base = declarative_base()
-
-
-class Result(BaseModel):
-    """Network search result."""
-
-    headers: Dict[str, List[str]] = {
-        "full": [
-            "Query",
-            "Device",
-            "Interface",
-            "IP",
-            "VRF",
-            "ACL in",
-            "ACL out",
-            "Shutdown",
-            "Description",
-        ]
-    }
-    data: List[Dict[str, str]] = []
+from faddr.models import Base, Device, Interface, IPAddress, ModelFactory
+from faddr.schemas import DeviceSchema, Result
 
 
-class Device(Base):  # pylint: disable=too-few-public-methods
-    """ORM 'device' table data mapping."""
-
-    __tablename__ = "device"
-
-    id = Column(Integer, primary_key=True)
-    name = Column(String, index=True)
-    path = Column(String)
-    source = Column(String)
+model_factory = ModelFactory()
 
 
-class Interface(Base):  # pylint: disable=too-few-public-methods
-    """ORM 'interface' table data mapping."""
+def make_sa_object(sa_class, data):
+    """Create SQLAlchemy table object from provided data."""
+    sa_obj_data = {}
+    for key in sa_class.__table__.columns.keys():
+        sa_obj_data[key] = dict(data).get(key)
 
-    __tablename__ = "interface"
+    for (relative, sa_subclass_name) in dict(data).get("sa_mapping", {}).items():
+        sa_sub_class = model_factory.get(sa_subclass_name)
+        # Create one-to-many relatives
+        if isinstance(dict(data).get(relative), list):
+            sa_obj_data[relative] = []
+            for sub in dict(data).get(relative):
+                sa_obj_data[relative].append(make_sa_object(sa_sub_class, sub))
 
-    id = Column(Integer, primary_key=True)
-    name = Column(String, index=True)
-    parent_interface = Column(String)
-    unit = Column(String)
-    duplex = Column(String)
-    speed = Column(String)
-    description = Column(String, index=True)
-    is_disabled = Column(Boolean, default=False)
-    encapsulation = Column(String)
-    s_vlan = Column(Integer)
-    c_vlan = Column(Integer)
-    vrf = Column(String, index=True)
-    acl_in = Column(String, index=True)
-    acl_out = Column(String, index=True)
-
-    device_id = Column(Integer, ForeignKey("device.id"))
-
-
-class IP(Base):  # pylint: disable=too-few-public-methods
-    """ORM 'ip' table data mapping."""
-
-    __tablename__ = "ip"
-
-    id = Column(Integer, primary_key=True)
-    broadcast_address = Column(String)
-    compressed = Column(String)
-    exploded = Column(String)
-    hostmask = Column(String)
-    hosts = Column(Integer)
-    ip = Column(String, index=True)
-    is_link_local = Column(Boolean, default=False)
-    is_loopback = Column(Boolean, default=False)
-    is_multicast = Column(Boolean, default=False)
-    is_private = Column(Boolean, default=False)
-    is_reserved = Column(Boolean, default=False)
-    is_unspecified = Column(Boolean, default=False)
-    max_prefixlen = Column(Integer)
-    netmask = Column(String)
-    network = Column(String, index=True)
-    network_address = Column(String, index=True)
-    num_addresses = Column(Integer)
-    prefixlen = Column(Integer)
-    version = Column(Integer)
-    with_hostmask = Column(String)
-    with_netmask = Column(String)
-    with_prefixlen = Column(String)
-
-    interface_id = Column(Integer, ForeignKey("interface.id"))
+    return sa_class(**sa_obj_data)
 
 
 class Database:
@@ -146,7 +71,7 @@ class Database:
         return engine
 
     def new_revision(self, revision=None):
-        """Create new revision and return it."""
+        """Create new revision and IP it."""
         if revision:
             self.revision = revision
         else:
@@ -164,54 +89,13 @@ class Database:
     def insert_device(self, device_data):
         """Insert device data to database."""
 
-        device = Device(**device_data["info"])
+        device = make_sa_object(Device, DeviceSchema.parse_obj(device_data))
 
         with Session(self.engine) as session:
             session.add(device)
             session.commit()
-            device_id = device.id
 
-        for interface_name, interface_data in device_data.get("interfaces", {}).items():
-            self.insert_interface(device_id, interface_name, interface_data)
-
-        logger.debug(f"Inserted device: '{device_data['info']}'")
-
-    def insert_interface(self, device_id, name, data):
-        """Insert interface data to database."""
-
-        table_data = {}
-        for key in Interface.__table__.columns.keys():
-            table_data[key] = data.get(key)
-        table_data["name"] = name
-        table_data["device_id"] = device_id
-
-        interface = Interface(**table_data)
-
-        with Session(self.engine) as session:
-            session.add(interface)
-            session.commit()
-            interface_id = interface.id
-
-        for ip_address in data.get("ip", []):
-            self.insert_ip_address(interface_id, ip_address)
-
-        logger.debug(f"Interted interface {data}")
-
-    def insert_ip_address(self, interface_id, data):
-        """Insert ip_address data to database."""
-
-        table_data = {}
-        for key in IP.__table__.columns.keys():
-            table_data[key] = data.get(key)
-        table_data["interface_id"] = interface_id
-
-        ip_address = IP(**table_data)
-
-        with Session(self.engine) as session:
-            session.add(ip_address)
-            session.commit()
-
-        logger.debug(f"Inserted IP address {data}")
+        logger.debug(f"Inserted device: '{device_data['name']}'")
 
     def set_default(self):
         """Make current revision default one."""
@@ -284,7 +168,7 @@ class Database:
             select(
                 Device.name.label("device"),
                 Interface.name.label("interface"),
-                IP.with_prefixlen,
+                IPAddress.with_prefixlen,
                 Interface.vrf,
                 Interface.acl_in,
                 Interface.acl_out,
@@ -292,13 +176,13 @@ class Database:
                 Interface.description,
             )
             .where(
-                IP.network.in_(networks),
-                Interface.id == IP.interface_id,
+                IPAddress.network.in_(networks),
+                Interface.id == IPAddress.interface_id,
                 Device.id == Interface.device_id,
             )
             .order_by(Device.name)
             .order_by(Interface.name)
-            .order_by(IP.with_prefixlen)
+            .order_by(IPAddress.with_prefixlen)
         )
 
         with Session(self.engine) as session: