+++ faddr/faddr_db.py
@@ -4,6 +4,7 @@ import argparse
 import sys
 
 import ray
+from pydantic import ValidationError
 
 from faddr import __version__, logger
 from faddr.database import Database
@@ -48,13 +49,6 @@ def parse_cmd_args():
 @ray.remote
 def parse_config(config, profile=None, template_dir=None):
     """Parse provided configuration and return structured data."""
-    device = {
-        "info": {
-            "path": str(config["path"]),
-            "name": str(config["name"]),
-            "source": "rancid",
-        }
-    }
 
     logger.info(f'Parsing \'{config["name"]}\'')
     try:
@@ -63,25 +57,41 @@ def parse_config(config, profile=None, template_dir=None):
             profile=profile,
             template_dir=template_dir,
         )
-        data = parser.parse()
-        device.update(data)
+        device = parser.parse()
     except FaddrParserUnknownProfile:
         logger.warning(f"Unsupported content-type in '{config}'")
     except FaddrParserConfigFileAbsent:
         logger.warning(f"Config file absent: '{config}'")
     except FaddrParserConfigFileEmpty:
         logger.warning(f"Config file empty: '{config}'")
+    except ValidationError as err:
+        logger.warning(f"Failed to postprocess '{config}': {err.json()}")
+
+    device.update(
+        {
+            "path": str(config["path"]),
+            "name": str(config["name"]),
+            "source": "rancid",
+        }
+    )
+
     return device
 
 
 @ray.remote
 def store_in_db(database, device):
     """Insert device data to database."""
-    if len(device) < 2:
-        logger.warning(f'Device \'{device["info"]["name"]}\' data is empty, skipping')
+    data_fields = ("interfaces",)
+    device_have_data = False
+    for data_field in data_fields:
+        if len(device[data_field]) > 0:
+            device_have_data = True
+
+    if not device_have_data:
+        logger.warning(f'Device \'{device["name"]}\' data is empty, skipping')
         return False
 
-    logger.info(f'Inserting \'{device["info"]["name"]}\' info DB')
+    logger.info(f'Inserting \'{device["name"]}\' info DB')
     database.insert_device(device)
     return True